pyproject.toml
[project]
name = "aprendendo"
version = "0.1.0"
description = ""
authors = [
    {name = "Guilherme",email = "guilhermemmatos07@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "python-jose (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
